Added base funtionality to games controller
